+++ exporter/otlphttpexporter/otlp.go
@@ -25,7 +25,6 @@ import (
 	"net/url"
 	"runtime"
 	"strconv"
-	"strings"
 	"time"
 
 	"go.uber.org/zap"
@@ -34,10 +33,8 @@ import (
 
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/config"
-	"go.opentelemetry.io/collector/config/configgrpc"
 	"go.opentelemetry.io/collector/consumer/consumererror"
 	"go.opentelemetry.io/collector/exporter/exporterhelper"
-	"go.opentelemetry.io/collector/internal/middleware"
 	"go.opentelemetry.io/collector/model/otlpgrpc"
 	"go.opentelemetry.io/collector/model/pdata"
 )
@@ -88,14 +85,6 @@ func (e *exporter) start(_ context.Context, host component.Host) error {
 	client, err := e.config.HTTPClientSettings.ToClient(host.GetExtensions())
 	if err != nil {
 		return err
-	}
-
-	if e.config.Compression != "" {
-		if strings.ToLower(e.config.Compression) == configgrpc.CompressionGzip {
-			client.Transport = middleware.NewCompressRoundTripper(client.Transport)
-		} else {
-			return fmt.Errorf("unsupported compression type %q", e.config.Compression)
-		}
 	}
 	e.client = client
 	return nil