+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## 💡 Enhancements 💡
 
+- `confighttp`: add client-side compression support. (#4441)
+  - Each exporter should remove `compression` field if they have and should use `confighttp.HTTPClientSettings`
 -  Allow more zap logger configs: `disable_caller`, `disable_stacktrace`, `output_paths`, `error_output_paths`, `initial_fields` (#1048)
 - `configauth`: add ServerAuthenticator interfaces for HTTP receivers. (#4506)